reorganise python files
