+++ charging_stations_pipelines/mapping/stations.py
@@ -6,10 +6,10 @@ import pandas as pd
 from dateutil import parser
 from shapely.geometry import Point
 
-from models.address import Address
-from models.station import Station
-from utils.bna_functions import check_coordinates
-from utils.logging_utils import log
+from charging_stations_pipelines.models.address import Address
+from charging_stations_pipelines.models.station import Station
+from charging_stations_pipelines.utils.bna_functions import check_coordinates
+from charging_stations_pipelines.utils.logging_utils import log
 from geoalchemy2.shape import from_shape
