+++ charging_stations_pipelines/models/address.py
@@ -2,7 +2,7 @@ from sqlalchemy import Column, Date, ForeignKey, Integer, String
 from sqlalchemy.orm import relationship
 from sqlalchemy.types import Float
 
-from models import Base
+from charging_stations_pipelines.models import Base
 
 
 class Address(Base):
