+++ charging_stations_pipelines/pipelines/_gbgov.py
@@ -7,13 +7,13 @@ from typing import Dict, Optional
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session
 
-from mapping.charging import map_charging_ocm
-from mapping.stations import map_address_ocm, map_station_ocm
-from utils.logging_utils import log
-from utils.ocm_extractor import ocm_extractor
+from charging_stations_pipelines.mapping.charging import map_charging_gb
+from charging_stations_pipelines.mapping.stations import map_address_gb, map_station_gb
+from charging_stations_pipelines.utils.gb_receiver import get_gb_data
+from charging_stations_pipelines.utils.logging_utils import log
 
 
-class OcmPipeline:
+class GbPipeline:
     def __init__(self, country_code:str, config: configparser, session: Session, offline: bool = False):
         self.country_code = country_code
         self.config = config
@@ -23,22 +23,22 @@ class OcmPipeline:
 
     def _retrieve_data(self):
         data_dir: str = os.path.join(
-            pathlib.Path(__file__).parent.resolve(), "..", "data"
+            pathlib.Path(__file__).parent.resolve(), "../..", "data"
         )
         pathlib.Path(data_dir).mkdir(parents=True, exist_ok=True)
-        tmp_file_path = os.path.join(data_dir, self.config["OCM"]["filename"])
+        tmp_file_path = os.path.join(data_dir, self.config["GBGOV"]["filename"])
         if not self.offline:
-            ocm_extractor(tmp_file_path, self.country_code)
+            get_gb_data(tmp_file_path)
         with open(tmp_file_path, "r") as f:
             self.data = json.load(f)
 
     def run(self):
         self._retrieve_data()
         entry: Dict
-        for id, entry in self.data.items():
-            mapped_address = map_address_ocm(entry, None)
-            mapped_charging = map_charging_ocm(entry, None)
-            mapped_station = map_station_ocm(entry, self.country_code)
+        for entry in self.data.get("ChargeDevice", []):
+            mapped_address = map_address_gb(entry, None)
+            mapped_charging = map_charging_gb(entry, None)
+            mapped_station = map_station_gb(entry, self.country_code)
             mapped_station.address = mapped_address
             mapped_station.charging = mapped_charging
             self.session.add(mapped_station)
@@ -46,10 +46,10 @@ class OcmPipeline:
                 self.session.commit()
                 self.session.flush()
             except IntegrityError as e:
-                log.error(f"OCM-Entry exists already! Error: {e}")
+                log.error(f"GBGOV-Entry exists already! Error: {e}")
                 self.session.rollback()
                 continue
             except Exception as e:
-                log.error(f"OCM-Pipeline failed to run! Error: {e}")
-                print(e)
+                log.error(f"GB-Pipeline failed to run! Error: {e}")
                 self.session.rollback()
+
