+++ charging_stations_pipelines/pipelines/_osm.py
@@ -7,15 +7,14 @@ from typing import Dict, Optional
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session
 
-from mapping.charging import map_charging_gb
-from mapping.stations import map_address_gb, map_station_gb
-from services.excel_file_loader_service import ExcelFileLoaderService
-from utils.gb_receiver import get_gb_data
-from utils.logging_utils import log
+from charging_stations_pipelines.mapping.charging import map_charging_osm
+from charging_stations_pipelines.mapping.stations import map_address_osm, map_station_osm
+from charging_stations_pipelines.utils.logging_utils import log
+from charging_stations_pipelines.utils.osm_receiver import get_osm_data
 
 
-class GbPipeline:
-    def __init__(self, country_code:str, config: configparser, session: Session, offline: bool = False):
+class OsmPipeline:
+    def __init__(self, country_code: str, config: configparser, session: Session, offline: bool = False):
         self.country_code = country_code
         self.config = config
         self.session = session
@@ -24,22 +23,22 @@ class GbPipeline:
 
     def _retrieve_data(self):
         data_dir: str = os.path.join(
-            pathlib.Path(__file__).parent.resolve(), "..", "data"
+            pathlib.Path(__file__).parent.resolve(), "../..", "data"
         )
         pathlib.Path(data_dir).mkdir(parents=True, exist_ok=True)
-        tmp_file_path = os.path.join(data_dir, self.config["GBGOV"]["filename"])
+        tmp_file_path = os.path.join(data_dir, self.config["OSM"]["filename"])
         if not self.offline:
-            get_gb_data(tmp_file_path)
+            get_osm_data(self.country_code, tmp_file_path)
         with open(tmp_file_path, "r") as f:
             self.data = json.load(f)
 
     def run(self):
         self._retrieve_data()
         entry: Dict
-        for entry in self.data.get("ChargeDevice", []):
-            mapped_address = map_address_gb(entry, None)
-            mapped_charging = map_charging_gb(entry, None)
-            mapped_station = map_station_gb(entry, self.country_code)
+        for entry in self.data.get("elements", []):
+            mapped_address = map_address_osm(entry, None)
+            mapped_charging = map_charging_osm(entry, None)
+            mapped_station = map_station_osm(entry, self.country_code)
             mapped_station.address = mapped_address
             mapped_station.charging = mapped_charging
             self.session.add(mapped_station)
@@ -47,10 +46,9 @@ class GbPipeline:
                 self.session.commit()
                 self.session.flush()
             except IntegrityError as e:
-                log.error(f"GBGOV-Entry exists already! Error: {e}")
+                log.error(f"OSM-Entry exists already! Error: {e}")
                 self.session.rollback()
                 continue
             except Exception as e:
-                log.error(f"GB-Pipeline failed to run! Error: {e}")
+                log.error(f"OSM-Pipeline failed to run! Error: {e}")
                 self.session.rollback()
-
