+++ settings.py
@@ -1,4 +1,3 @@
-import logging
 import logging.config
 import os
 import sys
